refactor: retire ecstore store api module
