+++ crates/ecstore/src/bucket/replication/replication_pool.rs
@@ -27,8 +27,8 @@ use crate::bucket::replication::replication_state::ReplicationStats;
 use crate::config::com::{read_config, save_config};
 use crate::disk::BUCKET_META_PREFIX;
 use crate::error::Error as EcstoreError;
+use crate::object_api::{ObjectInfo, ObjectOptions};
 use crate::storage_api_contracts::EcstoreObjectIO;
-use crate::store_api::{ObjectInfo, ObjectOptions};
 use lazy_static::lazy_static;
 use rustfs_filemeta::MrfOpKind;
 use rustfs_filemeta::MrfReplicateEntry;
