+++ crates/ecstore/src/bucket/replication/replication_resyncer.rs
@@ -31,13 +31,13 @@ use crate::error::{Error, Result, is_err_object_not_found, is_err_version_not_fo
 use crate::event_notification::{EventArgs, send_event};
 use crate::global::GLOBAL_LocalNodeName;
 use crate::global::get_global_bucket_monitor;
-use crate::resolve_object_store_handle;
-use crate::set_disk::get_lock_acquire_timeout;
-use crate::storage_api_contracts::{EcstoreObjectIO, EcstoreObjectOperations};
-use crate::store_api::{
+use crate::object_api::{
     GetObjectReader, ListObjectVersionsInfo, ListObjectsV2Info, ObjectInfo, ObjectInfoOrErr, ObjectOptions, PutObjReader,
     WalkOptions,
 };
+use crate::resolve_object_store_handle;
+use crate::set_disk::get_lock_acquire_timeout;
+use crate::storage_api_contracts::{EcstoreObjectIO, EcstoreObjectOperations};
 use aws_sdk_s3::error::{ProvideErrorMetadata, SdkError};
 use aws_sdk_s3::operation::head_object::{HeadObjectError, HeadObjectOutput};
 use aws_sdk_s3::primitives::ByteStream;
