+++ crates/ecstore/src/client/object_api_utils.rs
@@ -25,7 +25,7 @@ use std::{collections::HashMap, io::Cursor, sync::Arc};
 use tokio::io::BufReader;
 
 use crate::error::ErrorResponse;
-use crate::store_api::{GetObjectReader, ObjectInfo, ObjectOptions};
+use crate::object_api::{GetObjectReader, ObjectInfo, ObjectOptions};
 use rustfs_filemeta::ObjectPartInfo;
 use rustfs_rio::HashReader;
 use rustfs_storage_api::HTTPRangeSpec;
