--- crates/ecstore/src/object_api.rs
@@ -1,4 +1,0 @@
-pub use crate::store_api::{
-    DeletedObject, GetObjectReader, ListObjectVersionsInfo, ListObjectsInfo, ListObjectsV2Info, ObjectInfo, ObjectInfoOrErr,
-    ObjectOptions, ObjectToDelete, PutObjReader, WalkOptions,
-};
