+++ crates/ecstore/src/store_list_objects.rs
@@ -21,10 +21,10 @@ use crate::disk::{DiskInfo, DiskStore};
 use crate::error::{
     Error, Result, StorageError, is_all_not_found, is_all_volume_not_found, is_err_bucket_not_found, to_object_err,
 };
+use crate::object_api::{ListObjectVersionsInfo, ListObjectsInfo, ObjectInfo, ObjectInfoOrErr, ObjectOptions, WalkOptions};
 use crate::set_disk::SetDisks;
-use crate::store_api::{ListObjectVersionsInfo, ListObjectsInfo, ObjectInfo, ObjectInfoOrErr, ObjectOptions, WalkOptions};
 use crate::store_utils::is_reserved_or_invalid_bucket;
-use crate::{store::ECStore, store_api::ListObjectsV2Info};
+use crate::{object_api::ListObjectsV2Info, store::ECStore};
 use futures::future::join_all;
 use rand::seq::SliceRandom;
 use rustfs_filemeta::{
